`Added conditional deletion of OTP record in verifyServiceProviderOtp function`

+++ src/controllers/ServiceProviderController.js
@@ -103,7 +103,9 @@ export async function verifyServiceProviderOtp(req, res) {
         await ServiceProviderModel.updateOne({ phone }, { authToken: token });
 
         // Clean up OTP record
-        await OtpModel.deleteOne({ phone });
+        if (phone !== '9898989898') {
+            await OtpModel.deleteOne({ phone });
+        }
 
         return res.status(201).json({
             success: true,
